backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
boto3==1.35.0
python-multipart==0.0.12
email-validator==2.2.0
# Pipeline — task orchestration
celery==5.4.0
redis==5.2.0
psycopg2-binary==2.9.10
